requirements.txt
django
django-heroku
mysqlclient
djangorestframework
setuptools
ConfigParser
xlwt
Pillow
gunicorn
mysql-connector-python
gevent --pre
auto-py-to-exe
django-cors-headers
django-rest-knox
reportlab

